+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Changelog
 =========
 
+## Unreleased
+### Added
+- Webhook controller and reference in CP
+
 ## 1.0.0 - 2018-05-23
 Initial GA release.